requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
werkzeug==3.0.4
pyjwt==2.9.0